+++ ch11/code/platform.mem.h
@@ -3,7 +3,7 @@
 #include "db.h"
 #include "errors.h"
 
-result_t palmem_allocate_pages(void **p, uint32_t pages);
+result_t palmem_allocate_pages(void **p, uint64_t pages);
 
 result_t palmem_free_page(void **p);
